+++ docs/api/puppeteer.browser.extensions.md
@@ -4,16 +4,16 @@ sidebar_label: Browser.extensions
 
 # Browser.extensions() method
 
-Get a list of the installed extensions in the browser.
+Get a map with id as keys and extension as value of the installed extensions in the browser.
 
 ### Signature
 
 ```typescript
 class Browser {
-  abstract extensions(): Promise<Extension[]>;
+  abstract extensions(): Promise<Map<string, Extension>>;
 }
 ```
 
 **Returns:**
 
-Promise&lt;[Extension](./puppeteer.extension.md)\[\]&gt;
+Promise&lt;Map&lt;string, [Extension](./puppeteer.extension.md)&gt;&gt;
