+++ packages/puppeteer-core/src/bidi/Browser.ts
@@ -377,15 +377,7 @@ export class BidiBrowser extends Browser {
     return this.#networkEnabled;
   }
 
-  override extensions(): Promise<Extension[]> {
-    throw new Error(
-      'Method not yet implemented for BiDi. Use CDP to interact with Extensions.',
-    );
-  }
-
-  override getExtensionById(_extensionId: string): Promise<Extension | null> {
-    throw new Error(
-      'Method not yet implemented for BiDi. Use CDP to interact with Extensions.',
-    );
+  override extensions(): Promise<Map<string, Extension>> {
+    throw new UnsupportedOperation();
   }
 }
