+++ test/TestExpectations.json
@@ -1879,104 +1879,6 @@
     "comment": "TODO: add a comment explaining why this expectation is required (include links to issues)"
   },
   {
-    "testIdPattern": "[extension.spec] Extensions should capture console logs from extension pages",
-    "platforms": ["linux"],
-    "parameters": ["cdp", "chrome", "chrome-headless-shell"],
-    "expectations": ["FAIL"],
-    "comment": "No extensions in chrome-headless-shell"
-  },
-  {
-    "testIdPattern": "[extension.spec] Extensions should capture console logs from extension pages",
-    "platforms": ["linux"],
-    "parameters": ["chrome", "headless", "webDriverBiDi"],
-    "expectations": ["FAIL"],
-    "comment": "No extensions in WebDriver BiDi so far"
-  },
-  {
-    "testIdPattern": "[extension.spec] Extensions should capture console logs from extension workers",
-    "platforms": ["linux"],
-    "parameters": ["cdp", "chrome", "chrome-headless-shell"],
-    "expectations": ["FAIL"],
-    "comment": "No extensions in chrome-headless-shell"
-  },
-  {
-    "testIdPattern": "[extension.spec] Extensions should capture console logs from extension workers",
-    "platforms": ["linux"],
-    "parameters": ["chrome", "headless", "webDriverBiDi"],
-    "expectations": ["FAIL"],
-    "comment": "No extensions in WebDriver BiDi so far"
-  },
-  {
-    "testIdPattern": "[extension.spec] Extensions should list extension pages",
-    "platforms": ["linux"],
-    "parameters": ["cdp", "chrome", "chrome-headless-shell"],
-    "expectations": ["FAIL"],
-    "comment": "No extensions in chrome-headless-shell"
-  },
-  {
-    "testIdPattern": "[extension.spec] Extensions should list extension pages",
-    "platforms": ["linux"],
-    "parameters": ["chrome", "headless", "webDriverBiDi"],
-    "expectations": ["FAIL"],
-    "comment": "No extensions in WebDriver BiDi so far"
-  },
-  {
-    "testIdPattern": "[extension.spec] Extensions should list extension workers",
-    "platforms": ["linux"],
-    "parameters": ["cdp", "chrome", "chrome-headless-shell"],
-    "expectations": ["FAIL"],
-    "comment": "No extensions in chrome-headless-shell"
-  },
-  {
-    "testIdPattern": "[extension.spec] Extensions should list extension workers",
-    "platforms": ["linux"],
-    "parameters": ["chrome", "headless", "webDriverBiDi"],
-    "expectations": ["FAIL"],
-    "comment": "No extensions in WebDriver BiDi so far"
-  },
-  {
-    "testIdPattern": "[extension.spec] Extensions should list extensions and their properties",
-    "platforms": ["linux"],
-    "parameters": ["cdp", "chrome", "chrome-headless-shell"],
-    "expectations": ["FAIL"],
-    "comment": "No extensions in chrome-headless-shell"
-  },
-  {
-    "testIdPattern": "[extension.spec] Extensions should list extensions and their properties",
-    "platforms": ["linux"],
-    "parameters": ["chrome", "headless", "webDriverBiDi"],
-    "expectations": ["FAIL"],
-    "comment": "No extensions in WebDriver BiDi so far"
-  },
-  {
-    "testIdPattern": "[extension.spec] Extensions should remove extension from list after uninstall",
-    "platforms": ["linux"],
-    "parameters": ["cdp", "chrome", "chrome-headless-shell"],
-    "expectations": ["FAIL"],
-    "comment": "No extensions in chrome-headless-shell"
-  },
-  {
-    "testIdPattern": "[extension.spec] Extensions should remove extension from list after uninstall",
-    "platforms": ["linux"],
-    "parameters": ["chrome", "headless", "webDriverBiDi"],
-    "expectations": ["FAIL"],
-    "comment": "No extensions in WebDriver BiDi so far"
-  },
-  {
-    "testIdPattern": "[extension.spec] Extensions should trigger extension action",
-    "platforms": ["linux"],
-    "parameters": ["cdp", "chrome", "chrome-headless-shell"],
-    "expectations": ["FAIL"],
-    "comment": "No extensions in chrome-headless-shell"
-  },
-  {
-    "testIdPattern": "[extension.spec] Extensions should trigger extension action",
-    "platforms": ["linux"],
-    "parameters": ["chrome", "headless", "webDriverBiDi"],
-    "expectations": ["FAIL"],
-    "comment": "No extensions in WebDriver BiDi so far"
-  },
-  {
     "testIdPattern": "[mouse.spec] Mouse should select the text with mouse",
     "platforms": ["win32"],
     "parameters": ["cdp", "chrome", "chrome-headless-shell"],
