+++ test/assets/extension-with-page/background.js
@@ -1,1 +1,3 @@
-console.log("example text");
+chrome.action.onClicked.addListener((tab) => {
+  console.log("example text");
+});
